Unsatisfying solution to git commit flyspell mode problem.
I ended up adding fundamental mode to the modes for which flyspell-mode is automatically enabled, but probably there is a reason I didn't do this before ... what I really want is for git commit to use text-mode, but I couldnt' figure out how to make this work.  It seems that vc-git reverts to fundamental-mode, even if I set text-mode explicitly.