Fix 'Exec' path for xmonad, which was broken by r219.
In 'r219' I changed the path from /home/collins/... to ~/... to make
it more robust, but apparently I didn't actually test it, and it turns
out the ~-path isn't supported.  See comments in

  dot.local/share/applications/local-xmonad-windowmanager-provider.desktop

for discussion.  The easy solution is that gnome-session looks in PATH
for the command, so just use 'Exec=xmonad' :P