Make compilation less disruptive.
My old "shrink the compilation window in all cases and delete the
compilation window on success" was very disruptive when the frame was
already split.  So, instead, now we restore the old buffer which the
compilation buffer kicked out on success.  If the frame was not
already split, then the split remains, but a 'C-x 1' fixes this.

Also, make compilation stop at the first error.  Nice, except I can't
figure out how to not stop at warnings.