ZSH: automatically merge shell history.
I used to avoid this, because it makes `<up>` and `<down>`
unintuitive, but here I restrict them to only step through the local
history. The `C-r` and `C-s` still search the global history, which is
now shared automatically, not just when I run `nc:mergehist` (which I
removed). To step through the global history, use `{M,C}-{p,n}`.