tmux: reload `SSH_AUTH_SOCK` when starting or reconnecting a tmux session.
According to the docs, `SSH_AUTH_SOCK` is part of the default value
for `update-environment`, so I don't know why it wasn't part of my
config (which is based on a config I got from Michael, which is based
on config he got from who knows where).

NB: changes to the tmux config are not respected by *new* sessions if
the tmux server is running. You have to manually reload the conf. This
is bound to `C-o r` in my conf.