[Xmonad] Use Stack to build .xmonad/xmonad.hs
This is a lot more complicated than I would like. I created a Cabal
package, `my-xmonad.cabal`, following
http://sitr.us/2018/05/13/build-xmonad-with-stack.html, and a
`.xmonad/build` script that rebuilds that package when `xmonad
--recompile` is run.

Part of this change is to make the entire `~/.xmonad` dir a link to
`v/conf/dot.xmonad`, instead of linking individual files like I had in
the past.