[Emacs] Move mode list from "mode line" to "header line"
And change how buffer names are displayed.

A recurring problem I have is that there is too much stuff in my mode
line and most of it gets cut off. So, I moved the mode list to a new
header line. This freed up a lot of space, so I changed the buffer
display to just be the file name (with minimal uniquification), and
added the full base path separately.

I implemented the full base path to replace "$HOME" with "~", which
helps in most cases, but I might also want to implement a length limit
on the base path, truncating a prefix when it's too long. The tricky
part about doing that correctly is that I'd like the max length to
depend on the available space in the mode line. The mode line docs
mentioned a function that renders a mode line and returns it as a
string, so I could probably use that to compute the available space.

It seems the buffer name gets artificially padded with spaces when
it's short, which isn't ideal, but also not a big deal.