[xmonad] Switch the keyboard remapping stuff back
Just had a Chesterton's Fence experience: running the hooks every
minute in cron seemed stupid, and I didn't remember why I had done it
that way, so I switched to the "better" solution of just running the
hooks once at startup. However, the xmodmap changes don't persist
across suspend and resume, which was the whole reason for the cron
solution in the first place. It's also possible to use hooks in
/etc/pm/sleep.d/, but that's much more complex than just using cron,
since those scripts run as root and need logic to restrict them to
only running on wake from suspend (altho it would be harmless to just
run my hacks all the time, including pointlessly when entering suspend
or hibernate).