[xmonad] Tweak M-tab mapping to not include other monitors
Before, when I cycled recent workspaces with M-tab and M-`, it would
first switch focus to the other monitors, since they are most
recent. This was never what I wanted, but I couldn't find a built-in
function that did what I wanted, so I left it this way for a long
time.