[Emacs] Fix buggy nc:custom-set-variable macro
I ran into a problem where using `when nil` to comment an
`nc:custom-set-variable` call resulted in the variable still getting
set. I.e. when I did something like this:

    (when nil
    (nc:custom-set-variable ...)
    )

Seems the problem was the my macro definition was not sufficiently
quoted, and so the variable was getting set at macro expansion time :/

Also, fix a typo in the name of an existing custom-set variable.