[Emacs] Add copilot completions
Finding interaction between copilot and company confusing with tab. I
had it so that sometimes tab completes copilot, sometimes not. But
pressing <right> always works for copilot, and pressing <enter> always
works for company, and sometimes both are available, so it seems fine
to make them different keys. In fact, might be better to have
tab *not* complete copilot, in case I want to tab to force company
completions to pop up. In this version tab does not complete copilot.