[Emacs] Configure ediff
Change defaults so that diffs are side-by-side, and control panel is in same Os
window as diffs.

I can't believe I only today got around to figuring out basic ediff usage. I
know I've entered ediff a few times over the years, but always got really
confused and quit. It seems the key problem is that you need to enter commands
in the "control panel", not the diff buffers, and it opens the control panel in
another Os window by default :/